update SttpMockServerClient example
