+++ build.sbt
@@ -1484,7 +1484,8 @@ lazy val examples: ProjectMatrix = (projectMatrix in file("examples"))
       "com.softwaremill.sttp.client3" %% "async-http-client-backend-cats" % Versions.sttp,
       "com.softwaremill.sttp.apispec" %% "asyncapi-circe-yaml" % Versions.sttpApispec,
       "com.pauldijou" %% "jwt-circe" % Versions.jwtScala,
-      scalaTest.value % Test
+      scalaTest.value % Test,
+      "org.mock-server" % "mockserver-netty-no-dependencies" % Versions.mockServer % Test
     ),
     libraryDependencies ++= loggerDependencies,
     publishArtifact := false
